Separated roomcell, implemented sidebar.fxml

+++ client/src/main/java/nl/tudelft/oopp/group39/room/controller/RoomReservationController.java
@@ -157,7 +157,7 @@ public class RoomReservationController extends MainSceneController {
      * @throws JsonProcessingException when there is something wrong with processing
      */
     public List<Integer> getBookedTimes(String date) throws JsonProcessingException {
-        String bookings = ServerCommunication.getBookings((int) room.getId(), date);
+        String bookings = ServerCommunication.getBookings(room.getId(), date);
         System.out.println(bookings);
         ArrayNode body = (ArrayNode) mapper.readTree(bookings).get("body");
         String bookingString = mapper.writeValueAsString(body);
